+++ cyberdrop_dl/managers/download_speed_manager.py
